Add interfaces and refactor some classes
