+++ app/src/main/java/ru/internetcloud/strava/data/common/ErrorMessageConverter.kt
@@ -1,30 +1,7 @@
 package ru.internetcloud.strava.data.common
 
-import android.app.Application
-import java.net.SocketTimeoutException
-import java.net.UnknownHostException
-import ru.internetcloud.strava.R
-import ru.internetcloud.strava.presentation.util.addLine
+interface ErrorMessageConverter {
+    fun getMessageToException(exception: Exception): String
 
-class ErrorMessageConverter(
-    private var applicaton: Application
-) {
-    fun getMessageToException(exception: Exception): String {
-        return if (exception is SocketTimeoutException || exception is UnknownHostException) {
-            applicaton.getString(R.string.no_internet_connection).addLine(exception.message.toString())
-        } else {
-            applicaton.getString(R.string.unknown_network_error).addLine(exception.message.toString())
-        }
-    }
-
-    fun getMessageToHTTPCode(code: Int): String {
-        val messageResId = when (code) {
-            400 -> R.string.bad_request_error
-            401 -> R.string.unauthorized_error
-            403 -> R.string.forbidden_error
-            500 -> R.string.internal_server_error
-            else -> R.string.unknown_network_error
-        }
-        return applicaton.getString(messageResId).addLine(applicaton.getString(R.string.http_error_code, code))
-    }
+    fun getMessageToHTTPCode(code: Int): String
 }
