+++ app/src/main/java/ru/internetcloud/strava/domain/common/model/SportTypeKeeperImpl.kt
@@ -1,0 +1,15 @@
+package ru.internetcloud.strava.domain.common.model
+
+class SportTypeKeeperImpl : SportTypeKeeper {
+    override fun getSportTypes(): List<String> {
+        return listOf(
+            "AlpineSki", "BackcountrySki", "Badminton", "Canoeing", "Crossfit", "EBikeRide",
+            "Elliptical", "EMountainBikeRide", "Golf", "GravelRide", "Handcycle", "HighIntensityIntervalTraining",
+            "Hike", "IceSkate", "InlineSkate", "Kayaking", "Kitesurf", "MountainBikeRide", "NordicSki",
+            "Pickleball", "Pilates", "Racquetball", "Ride", "RockClimbing", "RollerSki", "Rowing", "Run",
+            "Sail", "Skateboard", "Snowboard", "Snowshoe", "Soccer", "Squash", "StairStepper", "StandUpPaddling",
+            "Surfing", "Swim", "TableTennis", "Tennis", "TrailRun", "Velomobile", "VirtualRide", "VirtualRow",
+            "VirtualRun", "Walk", "WeightTraining", "Wheelchair", "Windsurf", "Workout", "Yoga"
+        )
+    }
+}
