+++ app/src/main/java/ru/internetcloud/strava/presentation/common/compose/ListProgressIndicator.kt
