+++ app/src/main/java/ru/internetcloud/strava/presentation/training/list/TrainingListScreen.kt
@@ -120,4 +120,3 @@ fun TrainingListScreen(
         }
     }
 }
-
